further dividing estimate.py into algorithms.py and dbaccess.py
